fix get_maximum_output_size overflow on 32-bit targets
Cast input_len to u64 before multiplying by 110, avoiding overflow on
32-bit targets (e.g. wasm32) where `input_len * 110` overflows usize
when input_len > 2^32 / 110 ≈ 39MB.

+++ src/block/compress.rs
@@ -586,7 +586,7 @@ fn init_dict<T: HashTable>(dict: &mut T, dict_data: &mut &[u8]) {
 /// Can be used to preallocate capacity on the output vector
 #[inline]
 pub const fn get_maximum_output_size(input_len: usize) -> usize {
-    16 + 4 + (input_len * 110 / 100)
+    16 + 4 + (input_len as u64 * 110 / 100) as usize
 }
 
 /// Compress all bytes of `input` into `output`.
